refactor(test): move common fixtures to root conftest file
