+++ tests/auth/test_register.py
@@ -5,7 +5,7 @@ from flask import Flask
 from flask.testing import FlaskClient
 
 from aphorism.apps.auth.model import User
-from tests.user.conftest import register_user
+from tests.conftest import register_user
 
 
 def test_register_validation(client: FlaskClient) -> None:
